# backend/nodejs/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Beauty Salon Management API
  description: |
    API completa para gerenciamento de salão de beleza, incluindo clientes, serviços, funcionários e agendamentos.
    
    ## Funcionalidades
    - Gerenciamento de clientes
    - Catálogo de serviços
    - Gestão de funcionários
    - Sistema de agendamentos
    - Monitoramento e métricas
    
    ## Autenticação
    Esta versão da API não implementa autenticação. Em produção, considere implementar JWT ou OAuth2.
    
    ## Rate Limiting
    A API implementa rate limiting de 100 requisições por 15 minutos por IP.
  version: 1.0.0
  contact:
    name: Beauty Salon Team
    email: dev@beautysalon.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:3000
    description: Servidor de desenvolvimento
  - url: https://api.beautysalon.com
    description: Servidor de produção
...
paths:
...
  /api/services/category/{category}/active:
    get:
      tags: [services]
      summary: Listar serviços ativos por categoria
      description: Retorna uma lista de serviços ativos filtrados por categoria
      operationId: getActiveServicesByCategory
      parameters:
        - name: category
          in: path
          required: true
          description: Categoria dos serviços
          schema:
            type: string
            example: "Hair"
      responses:
        '200':
          description: Lista de serviços ativos da categoria retornada com sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Service'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/services/{id}:
    put:
      tags: [services]
      summary: Atualizar serviço
      description: Atualiza os dados de um serviço existente
      operationId: updateService
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceInput'
      responses:
        '200':
          description: Serviço atualizado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Service'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    delete:
      tags: [services]
      summary: Excluir serviço
      description: Remove um serviço do sistema
      operationId: deleteService
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '204':
          description: Serviço excluído com sucesso
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/services:
    get:
      tags: [services]
      summary: Listar todos os serviços
      description: Retorna uma lista de todos os serviços disponíveis
      operationId: getServices
      responses:
        '200':
          description: Lista de serviços retornada com sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Service'
        '500':
          $ref: '#/components/responses/InternalServerError'
    
    post:
      tags: [services]
      summary: Criar novo serviço
      description: Cadastra um novo serviço no sistema
      operationId: createService
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ServiceInput'
      responses:
        '201':
          description: Serviço criado com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Service'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      description: ID único do recurso
      schema:
        type: string
        format: uuid
  schemas:
...
    Service:
      type: object
      required: [id, name, price, duration]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          example: "Corte Feminino"
        description:
          type: string
          example: "Corte de cabelo feminino com lavagem"
        price:
          type: number
          format: double
          example: 50.00
        duration:
          type: integer
          example: 60
        category:
          type: string
          example: "Cabelo"
        isActive:
          type: boolean
          example: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    ServiceInput:
      type: object
      required: [name, price, duration]
      properties:
        name:
          type: string
          example: "Corte Feminino"
        description:
          type: string
          example: "Corte de cabelo feminino com lavagem"
        price:
          type: number
          format: double
          example: 50.00
        duration:
          type: integer
          example: 60
        category:
          type: string
          example: "Cabelo"
        isActive:
          type: boolean
          example: true
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          example: "Validation failed"
        message:
          type: string
          example: "Required field 'name' is missing"
        timestamp:
          type: string
          format: date-time
  responses:
    BadRequest:
      description: Requisição inválida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    InternalServerError:
      description: Erro interno do servidor
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'